+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # CHANGELOG
 https://keepachangelog.com/en/1.0.0/
 
+## [1.22.2] - 2023-02-25
+- `ComparingStringsWithoutStringComparison`: Don't trigger when the same symbol is referenced in both operands
+- `SwitchDoesNotHandleAllEnumOptions`: When a `default` clause is present, the analyzer will become an INFO diagnostic rather than a WARN
+
 ## [1.22.1] - 2023-02-21
 - `ThreadSleepInAsyncMethod`: Fixed a `NullReferenceException` if symbols weren't loaded